Added lirc-gpio driver.
From Matthias Hoelling via the linux-sunxi list:

This patch enables transmitting and receiving IR signals via the LIRC socket on GPIO.
The LIRC_GPIO_ID_STRING parameter binds it to an Allwinner A1X chip and/or the
gpio-sunxi module.  If the string is changed, it may also work on other platforms.
Parameters can be adjusted in sysfs after module is loaded, so different transmitters
on different pins can be used (locations or different wavelength of diode), as well
as different receivers (bandpass frequencies of the IR receiver).

Only the Transmitter was thoroughly tested as this is what I am interested in (as
the Cubieboard has a IR RX, but no TX in hardware).

Signed-off-by: Matthias Hoelling <mhoel....@gmail.com>